releasetools: ota_from_target_files: modify boot partition update
In standard Android systems, the boot.img containing the kernel, the
device tree as well as the ramdisk is used. In our case, we chose to
have those elements separated and available under a FAT boot partition.
Therefore a raw write cannot be performed any more and the files need to
be copied manually. This patch also removes the update of the recovery
partition.

Signed-off-by: Gary Bisson <gary.bisson@boundarydevices.com>